fix: move pg_class tests to a separate class
